third_party/blink/web_tests/wpt_internal/atomic-move-auto/popover-preserve.html

<!DOCTYPE html>
<title>moveBefore should not close a popover</title>
<meta name="variant" content="?action=append">
<meta name="variant" content="?action=prepend">
<meta name="variant" content="?action=before">
<meta name="variant" content="?action=after">
<script src="resources/helper.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<section id="old_parent">
<div popover>
Popover
</div>
</section>
<section id="new_parent">
</section>
<script>
promise_test(async t => {
    const popover = document.querySelector("div[popover]");
    popover.showPopover();
    await new Promise(resolve => requestAnimationFrame(() => resolve()));
    assert_equals(document.querySelector(":popover-open"), popover);
    reparent(document.querySelector("#new_parent"), popover);
    assert_equals(document.querySelector(":popover-open"), popover);
}, "when reparenting an open popover, it shouldn't be closed automatically");
</script>
